feat: extract modality types, bus, wire protocol, and pipeline into pkg/modality
Moves the core modality abstractions (Module interface, Bus, WireMessage,
SubprocessConn, ChannelRegistry, Salience, ProcessSupervisor, TextModule)
into pkg/modality as a stdlib-only reusable package. Kernel files become
thin re-export wrappers via type aliases, preserving backward compatibility.

Layer A (types + interfaces) and Layer B (bus, wire, supervisor) are
fully extracted. Layer C (ledger integration, HTTP modules, serve wiring,
tools, HUD) remains in the kernel where it depends on kernel-specific types.

33 tests cover bus dispatch, wire serialization, channel registry,
salience scoring, event validation, and text module lifecycle.
